requirements.txt
asgiref==3.8.1
boto==2.49.0
botocore==1.38.9
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.3
defusedxml==0.7.1
dj-database-url==2.3.0
Django==5.2
django-autoslug==1.9.9
django-countries==7.6.1
django-debug-toolbar==5.2.0
django-model-utils==5.0.0
django-redis-cache==3.0.1
django-storages==1.14.6
django-storages-redux==1.3.3
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
oauthlib==3.2.2
packaging==25.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
redis==3.5.3
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.12.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
whitenoise==6.9.0
